+++ CHANGELOG.md
@@ -75,6 +75,7 @@ Support for fully symbolic contract addresses required some very extensive chang
 ## Changed
 
 - Removed sha3Crack which has been deprecated for keccakEqs
+- Abstraction-refinement for more complicated expressions such as MULMOD
 
 ## Added